requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
altair==5.1.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blip-ci==0.0.5
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
clip-interrogator==0.5.4
cmake==3.28.1
contourpy==1.1.1
cycler==0.12.0
diffusers==0.23.1
distlib==0.3.7
Django==4.2.7
django-environ==0.11.2
django-ninja==1.0.1
exceptiongroup==1.1.3
fairscale==0.4.4
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.0
fsspec==2023.9.2
ftfy==6.1.3
gdown==4.7.1
gradio==3.46.0
gradio_client==0.5.3
greenlet==1.1.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.0
invisible-watermark==0.2.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
opencv-python==4.9.0.80
orjson==3.9.7
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
platformdirs==4.0.0
protobuf==4.25.2
psutil==5.9.8
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rpds-py==0.10.3
safetensors==0.4.2
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
timm==0.4.12
tokenizers==0.15.1
toolz==0.12.0
torch==2.1.2
torchaudio
torchvision==0.16.2
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
virtualenv==20.24.7
wcwidth==0.2.13
websockets==11.0.3
zipp==3.17.0
